+++ packages/core/src/abstract-dialect/query-interface.js
@@ -85,6 +85,9 @@ export class AbstractQueryInterface extends AbstractQueryInterfaceTypeScript {
     // Postgres requires special SQL commands for ENUM/ENUM[]
     await this.ensureEnums(tableName, attributes, options, model);
 
+    // Snowflake requires special SQL commands for SEQUENCES
+    await this.ensureSequences(tableName, attributes, options);
+
     const modelTable = model?.table;
 
     if (!tableName.schema && (options.schema || modelTable?.schema)) {
@@ -858,6 +861,20 @@ export class AbstractQueryInterface extends AbstractQueryInterfaceTypeScript {
    * @private
    */
   ensureEnums() {
+    // noop by default
+  }
+
+  /**
+   * @private
+   */
+  ensureSequences() {
+    // noop by default
+  }
+
+  /**
+   * @private
+   */
+  getNextPrimaryKeyValue() {
     // noop by default
   }
 }