+++ packages/core/src/model-definition.ts
@@ -201,12 +201,13 @@ export class ModelDefinition<M extends Model = Model> {
 
     // TODO: deep freeze this.options
     // caution: mergeModelOptions mutates its first input
-    this.options = mergeModelOptions(
+    const validate = {} satisfies ModelOptions<M>['validate'];
+    this.options = mergeModelOptions<M>(
       // default options
       {
         noPrimaryKey: false,
         timestamps: true,
-        validate: {},
+        validate,
         freezeTableName: false,
         underscored: false,
         paranoid: false,
@@ -959,11 +960,11 @@ function banReferenceModel<T>(reference: T): T {
  * @param options
  * @param overrideOnConflict
  */
-export function mergeModelOptions(
-  existingModelOptions: ModelOptions,
-  options: ModelOptions,
+export function mergeModelOptions<M extends Model>(
+  existingModelOptions: ModelOptions<M>,
+  options: ModelOptions<M>,
   overrideOnConflict: boolean,
-): ModelOptions {
+): ModelOptions<M> {
   // merge-able: scopes, indexes
   for (const [optionName, optionValue] of Object.entries(options) as Array<
     [keyof ModelOptions, any]
@@ -981,7 +982,7 @@ export function mergeModelOptions(
           continue;
         }
 
-        if (!overrideOnConflict && subOptionName in existingModelOptions[optionName]!) {
+        if (!overrideOnConflict && subOptionName in existingModelOptions[optionName]) {
           throw new Error(
             `Trying to set the option ${optionName}[${JSON.stringify(subOptionName)}], but a value already exists.`,
           );
@@ -1008,7 +1009,8 @@ export function mergeModelOptions(
           : [existingHooks[hookType]];
 
         if (!Array.isArray(optionValue[hookType])) {
-          // @ts-expect-error -- typescript doesn't like this merge algorithm.
+          // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error -- became valid in TS 5.8
+          // @ts-ignore -- typescript doesn't like this merge algorithm.
           existingHooks[hookType] = [...existingHooksOfType, optionValue[hookType]];
         } else {
           // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error -- This error only occurs on TS 5.3+